refactor: Split indexes.rs into focused modules
Refactored the 1,097-line indexes.rs file into a well-organized module structure
for better maintainability and code organization.

## Module Structure

- **index_metadata.rs** (~80 lines): Types and helpers for index definitions
  - IndexMetadata struct
  - IndexData enum
  - normalize_index_name function
  - acquire_btree_lock helper
  - DISK_BACKED_THRESHOLD constant

- **index_operations.rs** (~240 lines): Query methods on IndexData
  - range_scan: Range predicate scanning
  - multi_lookup: IN predicate support
  - get: Point lookups
  - contains_key: Existence checks
  - iter/values: Full index scans

- **index_maintenance.rs** (~470 lines): CRUD operations for indexes
  - create_index: Index creation with disk/memory backend selection
  - drop_index/drop_indexes_for_table: Index removal
  - add_to_indexes_for_insert: Insert operations
  - update_indexes_for_update: Update operations
  - update_indexes_for_delete: Delete operations
  - rebuild_indexes: Bulk rebuild after compaction

- **index_manager.rs** (~390 lines): Core IndexManager coordination
  - IndexManager struct definition
  - new/with_config: Constructors
  - index_exists/get_index/get_index_data: Query methods
  - check_unique_constraints_for_insert: Constraint validation
  - enforce_memory_budget/spill_index_to_disk: Resource management
  - list_indexes: Index enumeration

- **mod.rs** (~20 lines): Public API re-exports

## Benefits

- Each module is now < 500 lines (target: 100-400 lines)
- Clear separation of concerns (metadata, queries, mutations, coordination)
- Easier to locate specific functionality
- Better testability for individual components
- Reduced cognitive load when working with any single piece

## Testing

- ✅ All 147 tests pass
- ✅ No new clippy warnings introduced
- ✅ Public API remains unchanged (backward compatible)
- ✅ cargo check passes without errors
